feat: added back button and translated most things to english to keep it consistent
I mean so far i had been using kind of a random mix of english and portuguese so i figured i better pick one and stick to it. There might still be portuguese being used in the console or something but i think the player will *hopefully* only see english for now.